+++ broker/src/broker/error.rs
@@ -1,14 +1,6 @@
-use std::error::Error;
-use std::fmt;
+use thiserror::Error;
 
 /// Broker has shut down.
-#[derive(Debug, Copy, Clone, PartialEq, Eq)]
+#[derive(Error, Debug, Copy, Clone, PartialEq, Eq)]
+#[error("broker shut down")]
 pub struct BrokerShutdown;
-
-impl fmt::Display for BrokerShutdown {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        f.write_str("broker shutdown")
-    }
-}
-
-impl Error for BrokerShutdown {}